Kind of Moles

+++ Bot.h
@@ -29,7 +29,7 @@ public:
     virtual TechTile chooseTechTile(const GameState& gs) = 0;
 
     virtual int8_t choosePlaceToSpade(const GameState& gs, int amount, const std::vector<int8_t>& possiblePos) = 0;
-    virtual int8_t choosePlaceForBridge(const GameState& gs, const std::vector<int8_t>& possiblePos) = 0;
+    virtual int choosePlaceForBridge(const GameState& gs, const std::vector<int>& possiblePos) = 0;
     virtual int8_t choosePlaceToBuildForFree(const GameState& gs, Building building, bool isNeutral, const std::vector<int8_t>& possiblePos) = 0;
     virtual int8_t chooseBuildingToConvertForFree(const GameState& gs, Building building, const std::vector<int8_t>& possiblePos) = 0;
